- Move commands into a common steps/functions - Update tests to leverage new functions instead of commands
