+++ tdrs-frontend/cypress/e2e/data-files/file_upload.js
@@ -1,5 +1,6 @@
 /* eslint-disable no-undef */
 import { Given, When, Then } from '@badeball/cypress-cucumber-preprocessor'
+import * as df from '../common-steps/data_files.js'
 
 //////////////////////// Generic User Steps ///////////////////////
 When('{string} uploads a file', (username) => {
@@ -21,14 +22,13 @@ Then('{string} can see the upload successful', (username) => {
 
 ///////////////////////// Admin Steps /////////////////////////
 Given('Admin Alex logs in', () => {
-  cy.restartAtHomePage().then(() => {
-    cy.login('admin-alex@acf.hhs.gov').then(() => {
-      cy.visit('/home')
-      cy.contains('FRA Data Files').should('exist')
-      cy.contains('Admin').should('exist')
-      cy.contains('Alerts').should('exist')
-      cy.contains('Grafana').should('exist')
-    })
+  df.restartAtHomePage()
+  cy.login('admin-alex@acf.hhs.gov').then(() => {
+    cy.visit('/home')
+    cy.contains('FRA Data Files').should('exist')
+    cy.contains('Admin').should('exist')
+    cy.contains('Alerts').should('exist')
+    cy.contains('Grafana').should('exist')
   })
 })
 
@@ -36,14 +36,13 @@ Given('Admin Alex logs in', () => {
 When('Admin Alex submits the TANF Report', () => {
   cy.visit('/data-files')
   cy.intercept('POST', '/v1/data_files/').as('dataFileSubmit')
-  cy.fillSttFyQ('New York', '2021', 'Q1', true).then(() => {
-    cy.uploadFile(
-      '#active-case-data',
-      '../tdrs-backend/tdpservice/parsers/test/data/small_correct_file.txt'
-    )
+  df.fillSttFyQ('New York', '2021', 'Q1', true)
+  df.uploadFile(
+    '#active-case-data',
+    '../tdrs-backend/tdpservice/parsers/test/data/small_correct_file.txt'
+  )
 
-    cy.get('button').contains('Submit Data Files').should('exist').click()
-  })
+  cy.get('button').contains('Submit Data Files').should('exist').click()
 })
 
 Then('Admin Alex sees the upload in TANF Submission History', () => {
@@ -58,30 +57,28 @@ Then('Admin Alex sees the upload in TANF Submission History', () => {
       const fileId = interception.response.body.id
 
       // Poll the API until the summary is populated
-      cy.waitForDataFileSummary(fileId).then(() => {
-        cy.get('button').contains('Submission History').click()
-        cy.validateSmallCorrectFile()
-      })
+      df.waitForDataFileSummary(fileId)
+      cy.get('button').contains('Submission History').click()
+      df.validateSmallCorrectFile()
     }
   })
 })
 
 Then('Admin Alex can download the TANF error report', () => {
-  cy.downloadErrorReport('2021-Q1-Active Case Data Error Report.xlsx')
+  df.downloadErrorReport('2021-Q1-Active Case Data Error Report.xlsx')
 })
 
 // SSP steps
 When('Admin Alex submits the SSP Report', () => {
   cy.visit('/data-files')
   cy.intercept('POST', '/v1/data_files/').as('dataFileSubmit')
-  cy.fillSttFyQ('Iowa', '2024', 'Q1', false).then(() => {
-    cy.uploadFile(
-      '#active-case-data',
-      '../tdrs-backend/tdpservice/parsers/test/data/small_ssp_section1.txt'
-    )
+  df.fillSttFyQ('Iowa', '2024', 'Q1', false)
+  df.uploadFile(
+    '#active-case-data',
+    '../tdrs-backend/tdpservice/parsers/test/data/small_ssp_section1.txt'
+  )
 
-    cy.get('button').contains('Submit Data Files').should('exist').click()
-  })
+  cy.get('button').contains('Submit Data Files').should('exist').click()
 })
 
 Then('Admin Alex sees the upload in SSP Submission History', () => {
@@ -96,29 +93,27 @@ Then('Admin Alex sees the upload in SSP Submission History', () => {
       const fileId = interception.response.body.id
 
       // Poll the API until the summary is populated
-      cy.waitForDataFileSummary(fileId).then(() => {
-        cy.get('button').contains('Submission History').click()
-        cy.validateSmallSSPFile()
-      })
+      df.waitForDataFileSummary(fileId)
+      cy.get('button').contains('Submission History').click()
+      df.validateSmallSSPFile()
     }
   })
 })
 
 Then('Admin Alex can download the SSP error report', () => {
-  cy.downloadErrorReport('2024-Q1-SSP Active Case Data Error Report.xlsx')
+  df.downloadErrorReport('2024-Q1-SSP Active Case Data Error Report.xlsx')
 })
 
 // FRA steps
 When('Admin Alex submits the Work Outcomes Report', () => {
   cy.visit('/fra-data-files')
   cy.intercept('POST', '/v1/data_files/').as('dataFileSubmit')
-  cy.fillSttFyQ('New York', '2024', 'Q2', false).then(() => {
-    cy.uploadFile(
-      '#fra-file-upload',
-      '../tdrs-backend/tdpservice/parsers/test/data/fra.csv'
-    )
-    cy.get('button').contains('Submit Report').should('exist').click()
-  })
+  df.fillSttFyQ('New York', '2024', 'Q2', false)
+  df.uploadFile(
+    '#fra-file-upload',
+    '../tdrs-backend/tdpservice/parsers/test/data/fra.csv'
+  )
+  cy.get('button').contains('Submit Report').should('exist').click()
 })
 
 Then('Admin Alex sees the upload in FRA Submission History', () => {
@@ -132,16 +127,15 @@ Then('Admin Alex sees the upload in FRA Submission History', () => {
       const fileId = interception.response.body.id
 
       // Poll the API until the summary is populated
-      cy.waitForDataFileSummary(fileId).then(() => {
-        cy.contains('Submission History').should('exist')
-        cy.validateFraCsv()
-      })
+      df.waitForDataFileSummary(fileId)
+      cy.contains('Submission History').should('exist')
+      df.validateFraCsv()
     }
   })
 })
 
 Then('Admin Alex can download the FRA error report', () => {
-  cy.downloadErrorReport(
+  df.downloadErrorReport(
     '2024-Q2-Work Outcomes of TANF Exiters Error Report.xlsx'
   )
 })
@@ -149,20 +143,19 @@ Then('Admin Alex can download the FRA error report', () => {
 
 /////////////////// FRA Data Analyst Steps ///////////////////
 Given('FRA Data Analyst Fred logs in', () => {
-  cy.restartAtHomePage().then(() => {
-    cy.login('fra-data-analyst-fred@teamraft.com').then(() => {
-      cy.visit('/home')
-      cy.contains('FRA Data Files').should('exist')
-    })
+  df.restartAtHomePage()
+  cy.login('fra-data-analyst-fred@teamraft.com').then(() => {
+    cy.visit('/home')
+    cy.contains('FRA Data Files').should('exist')
   })
 })
 
 // FRA steps
 When('FRA Data Analyst Fred submits the Work Outcomes Report', () => {
   cy.visit('/fra-data-files')
-  cy.fillFYQ('2024', 'Q2')
+  df.fillFYQ('2024', 'Q2')
   cy.intercept('POST', '/v1/data_files/').as('dataFileSubmit')
-  cy.uploadFile(
+  df.uploadFile(
     '#fra-file-upload',
     '../tdrs-backend/tdpservice/parsers/test/data/fra.csv'
   )
@@ -180,24 +173,23 @@ Then('FRA Data Analyst Fred sees the upload in Submission History', () => {
       const fileId = interception.response.body.id
 
       // Poll the API until the summary is populated
-      cy.waitForDataFileSummary(fileId).then(() => {
-        cy.contains('Submission History').should('exist')
-        cy.validateFraCsv()
-      })
+      df.waitForDataFileSummary(fileId)
+      cy.contains('Submission History').should('exist')
+      df.validateFraCsv()
     }
   })
 })
 
 Then('FRA Data Analyst Fred can download the FRA error report', () => {
-  cy.downloadErrorReport(
+  df.downloadErrorReport(
     '2024-Q2-Work Outcomes of TANF Exiters Error Report.xlsx'
   )
 })
 
 When('FRA Data Analyst Fred uploads incorrect file type', () => {
   cy.visit('/fra-data-files')
-  cy.fillFYQ('2024', 'Q2')
-  cy.uploadFile(
+  df.fillFYQ('2024', 'Q2')
+  df.uploadFile(
     '#fra-file-upload',
     '../tdrs-backend/tdpservice/parsers/test/data/small_correct_file.txt'
   )
@@ -213,9 +205,9 @@ Then('FRA Data Analyst Fred sees the incorrect file type error', () => {
 When('FRA Data Analyst Fred submits the TANF Report', () => {
   cy.visit('/data-files')
   cy.get(':nth-child(2) > .usa-radio__label').click()
-  cy.fillFYQ('2021', 'Q1')
+  df.fillFYQ('2021', 'Q1')
   cy.intercept('POST', '/v1/data_files/').as('dataFileSubmit')
-  cy.uploadFile(
+  df.uploadFile(
     '#active-case-data',
     '../tdrs-backend/tdpservice/parsers/test/data/small_correct_file.txt'
   )
@@ -234,25 +226,24 @@ Then('FRA Data Analyst Fred sees the upload in TANF Submission History', () => {
       const fileId = interception.response.body.id
 
       // Poll the API until the summary is populated
-      cy.waitForDataFileSummary(fileId).then(() => {
-        cy.get('button').contains('Submission History').click()
-        cy.validateSmallCorrectFile()
-      })
+      df.waitForDataFileSummary(fileId)
+      cy.get('button').contains('Submission History').click()
+      df.validateSmallCorrectFile()
     }
   })
 })
 
 Then('FRA Data Analyst Fred can download the TANF error report', () => {
-  cy.downloadErrorReport('2021-Q1-Active Case Data Error Report.xlsx')
+  df.downloadErrorReport('2021-Q1-Active Case Data Error Report.xlsx')
 })
 
 // SSP steps
 When('FRA Data Analyst Fred submits the SSP Report', () => {
   cy.visit('/data-files')
   cy.get(':nth-child(3) > .usa-radio__label').click()
-  cy.fillFYQ('2024', 'Q1')
+  df.fillFYQ('2024', 'Q1')
   cy.intercept('POST', '/v1/data_files/').as('dataFileSubmit')
-  cy.uploadFile(
+  df.uploadFile(
     '#active-case-data',
     '../tdrs-backend/tdpservice/parsers/test/data/small_ssp_section1.txt'
   )
@@ -271,16 +262,15 @@ Then('FRA Data Analyst Fred sees the upload in SSP Submission History', () => {
       const fileId = interception.response.body.id
 
       // Poll the API until the summary is populated
-      cy.waitForDataFileSummary(fileId).then(() => {
-        cy.get('button').contains('Submission History').click()
-        cy.validateSmallSSPFile()
-      })
+      df.waitForDataFileSummary(fileId)
+      cy.get('button').contains('Submission History').click()
+      df.validateSmallSSPFile()
     }
   })
 })
 
 Then('FRA Data Analyst Fred can download the SSP error report', () => {
-  cy.downloadErrorReport('2024-Q1-SSP Active Case Data Error Report.xlsx')
+  df.downloadErrorReport('2024-Q1-SSP Active Case Data Error Report.xlsx')
 })
 
 ///////////////////////////////////////////////////////////////
