+++ tdrs-frontend/cypress/support/commands.js
@@ -157,129 +157,3 @@ Cypress.Commands.add(
     return cy.get('@response')
   }
 )
-
-Cypress.Commands.add('restartAtHomePage', () => {
-  cy.clearCookie('sessionid')
-  cy.clearCookie('csrftoken')
-  cy.intercept('/v1/stts/alpha').as('getSttSearchList')
-  cy.visit('/')
-  cy.contains('Sign into TANF Data Portal', { timeout: 30000 })
-})
-
-Cypress.Commands.add('fillFYQ', (fiscal_year, quarter) => {
-  cy.get('#reportingYears').should('exist').select(fiscal_year)
-  cy.get('#quarter').should('exist').select(quarter)
-  cy.get('button').contains('Search').should('exist').click()
-  cy.wait(100)
-})
-
-Cypress.Commands.add('uploadFile', (file_input, file_path) => {
-  cy.get(file_input).selectFile(file_path, { action: 'drag-drop' })
-  cy.wait(500)
-})
-
-Cypress.Commands.add('validateSmallCorrectFile', () => {
-  cy.get('th').contains('small_correct_file.txt').should('exist')
-  cy.get('th').contains('1').should('exist')
-  cy.get('th').contains('Rejected').should('exist')
-  cy.get('th')
-    .contains('2021-Q1-Active Case Data Error Report.xlsx')
-    .should('exist')
-})
-
-Cypress.Commands.add('validateSmallSSPFile', () => {
-  cy.get('th').contains('small_ssp_section1.txt').should('exist')
-  cy.get('th').contains('1').should('exist')
-  cy.get('th').contains('5').should('exist')
-  cy.get('th').contains('Partially Accepted with Errors').should('exist')
-  cy.get('th')
-    .contains('2024-Q1-SSP Active Case Data Error Report.xlsx')
-    .should('exist')
-})
-
-Cypress.Commands.add('validateFraCsv', () => {
-  cy.get('td').contains('fra.csv').should('exist')
-  cy.get('td').contains('8').should('exist')
-  cy.get('td').contains('Partially Accepted with Errors').should('exist')
-  cy.get('td')
-    .contains('2024-Q2-Work Outcomes of TANF Exiters Error Report.xlsx')
-    .should('exist')
-})
-
-Cypress.Commands.add('downloadErrorReport', (error_report_name) => {
-  cy.get('button').contains(error_report_name).should('exist').click()
-  cy.readFile(`${Cypress.config('downloadsFolder')}/${error_report_name}`)
-})
-
-Cypress.Commands.add('fillSttFyQ', (stt, fy, q, isTanf) => {
-  cy.get('#stt')
-    .type(stt + '{enter}')
-    .then(() => {
-      if (isTanf) {
-        cy.get(':nth-child(2) > .usa-radio__label').click()
-      } else {
-        cy.get(':nth-child(3) > .usa-radio__label').click()
-      }
-      cy.get('#reportingYears').should('exist').select(fy)
-      cy.get('#quarter').should('exist').select(q)
-      cy.get('button').contains('Search').should('exist').click()
-      cy.wait(100)
-    })
-})
-
-Cypress.Commands.add('fillSttFyQNoProgramSelector', (stt, fy, q) => {
-  cy.get('#stt')
-    .type(stt + '{enter}')
-    .then(() => {
-      cy.get('#reportingYears').should('exist').select(fy)
-      cy.get('#quarter').should('exist').select(q)
-      cy.get('button').contains('Search').should('exist').click()
-      cy.wait(100)
-    })
-})
-
-Cypress.Commands.add(
-  'waitForDataFileSummary',
-  (fileId, maxAttempts = 60, interval = 2000) => {
-    // Function to check if summary exists and is populated
-    const checkSummary = (response) => {
-      return (
-        response &&
-        response.body &&
-        response.body.summary &&
-        Object.keys(response.body.summary).length > 0 &&
-        response.body.summary.status !== 'Pending'
-      )
-    }
-
-    const pollForProcessing = (attempt = 0) => {
-      // If we've exceeded max attempts, should we do anything else?
-      if (attempt >= maxAttempts) {
-        cy.log(
-          `Warning: Data file ${fileId} processing timeout after ${maxAttempts} attempts`
-        )
-        return cy.wrap({ id: fileId })
-      }
-
-      return cy
-        .request({
-          method: 'GET',
-          url: `${Cypress.env('apiUrl')}/data_files/${fileId}/`,
-          failOnStatusCode: false,
-        })
-        .then((response) => {
-          // If summary is populated, return the response
-          if (checkSummary(response)) {
-            return response
-          }
-
-          // Otherwise, wait and try again
-          cy.wait(interval)
-          return pollForProcessing(attempt + 1)
-        })
-    }
-
-    // Start polling
-    return pollForProcessing()
-  }
-)
